AVRCT CT: uint8_t is used for 2byte field
This causes the upper stack to report incorrect entries if the value is
large

Bug: b/34682784
Test: Check with a list > 4 items large and packet size big enough to
exceed uint8_t size

+++ stack/avrc/avrc_pars_ct.cc
@@ -163,7 +163,7 @@ static tAVRC_STS avrc_pars_browse_rsp(tAVRC_MSG_BROWSE* p_msg,
 
   /* read the pdu */
   BE_STREAM_TO_UINT8(pdu, p);
-  uint8_t pkt_len;
+  uint16_t pkt_len;
   /* read the entire packet len */
   BE_STREAM_TO_UINT16(pkt_len, p);
